refactor(hogql): materialized column handling
